jpa models updated to match deploy db. fixed bugs created when updating the models

+++ interfaces/src/main/java/ar/edu/itba/paw/persistence/AlbumDao.java
@@ -8,7 +8,7 @@ public interface AlbumDao extends CrudDao<Album> {
     List<Album> findByArtistId(long id);
     List<Album> findByTitleContaining(String sub);
 
-    boolean updateRating(long albumId, float newRating, int newRatingAmount);
+    boolean updateRating(long albumId, double newRating, int newRatingAmount);
     boolean hasUserReviewed(long userId, long albumId);
 
     boolean deleteReviewsFromAlbum(long albumId);
